regression: split the simulation framework from the suites which use it
