+++ test/sim/cocotb/arch_elf_entrypoint.py
@@ -1,7 +1,7 @@
 import os
 import cocotb
 
-from test.regression.cocotb import CocotbSimulation
+from test.sim.cocotb import CocotbSimulation
 from test.regression.test_arch_regression import run_arch_elf
